stream parallel

+++ src/main/java/com/homework/homework_school/controller/StudentController.java
@@ -85,7 +85,7 @@ public class StudentController {
     }
     //    средний возраст всех студентов
     @GetMapping("/avarageAgeOfStudents")
-    public int findAvarageAgeAllStudents() {
+    public double findAvarageAgeAllStudents() {
         return studentService.findAvarageAgeAllStudents();
     }
